Make standard rulesets above competitive
This is more of a personal opinion, but having ranked be the default ruleset doesn't make much sense to me. This game doesn't entirely revolve around the ranked ruleset for general people using it, and it shouldn't be the first option available, lest it confuses people who just want to play casually or with mods. This moves the standard rulesets up and comp movesets down one (the badlatro and sandbox are still at the bottom) and makes Blitz the default selected one.